Moving extensions types to a separate file
