+++ src/test/extensions/manifest.spec.ts
@@ -8,7 +8,7 @@ import * as refs from "../../extensions/refs";
 import { Config } from "../../config";
 import * as prompt from "../../prompt";
 import { FirebaseError } from "../../error";
-import { ParamType } from "../../extensions/extensionsApi";
+import { ParamType } from "../../extensions/types";
 
 /**
  * Returns a base Config with some extensions data.
